requirements.txt
flask~=2.3.3
Pillow~=10.0.1
APScheduler~=3.10.4